requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
mutagen>=1.47.0
pillow>=10.0.0
